+++ src/templates/inviteEmail.html
@@ -1,0 +1,82 @@
+<!doctype html>
+<html xmlns="http://www.w3.org/1999/xhtml">
+<head>
+    <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1">
+    <style type="text/css">
+        #outlook a { padding: 0; }
+        body { margin: 0; padding: 0; -webkit-text-size-adjust: 100%; -ms-text-size-adjust: 100%; background-color: #f9f9f9; }
+        table, td { border-collapse: collapse; mso-table-lspace: 0pt; mso-table-rspace: 0pt; }
+        img { border: 0; height: auto; line-height: 100%; outline: none; text-decoration: none; }
+        p { display: block; margin: 13px 0; }
+    </style>
+</head>
+<body style="background-color:#f9f9f9;">
+    <div style="background-color:#f9f9f9;">
+
+        <!-- Top border -->
+        <div style="background:#f9f9f9;margin:0px auto;max-width:600px;">
+            <table align="center" border="0" cellpadding="0" cellspacing="0" role="presentation" style="width:100%;">
+                <tbody><tr>
+                    <td style="border-bottom:#811429 solid 5px;font-size:0px;padding:20px 0;text-align:center;"></td>
+                </tr></tbody>
+            </table>
+        </div>
+
+        <!-- Main card -->
+        <div style="background:#fff;margin:0px auto;max-width:600px;">
+            <table align="center" border="0" cellpadding="0" cellspacing="0" role="presentation" style="background:#fff;width:100%;">
+                <tbody><tr>
+                    <td style="border:#dddddd solid 1px;border-top:0px;font-size:0px;padding:20px 0;text-align:center;">
+                        <table border="0" cellpadding="0" cellspacing="0" role="presentation" width="100%">
+
+                            <!-- Logo -->
+                            <tr>
+                                <td align="center" style="padding:30px 25px 10px;">
+                                    <img src="https://support.oc.edu/hc/theming_assets/01HZPDPGNX6721NE6REBJYW218" width="80%" style="display:block;border:0;outline:none;text-decoration:none;max-width:80%;margin:0 auto;padding:0 10%;" />
+                                </td>
+                            </tr>
+
+                            <!-- Heading -->
+                            <tr>
+                                <td align="center" style="padding:10px 25px 30px;">
+                                    <div style="font-family:'Helvetica Neue',Arial,sans-serif;font-size:28px;font-weight:bold;line-height:1;color:#811429;">
+                                        Welcome to {{Product}}
+                                    </div>
+                                </td>
+                            </tr>
+
+                            <!-- Message -->
+                            <tr>
+                                <td align="center" style="padding:10px 40px;">
+                                    <div style="font-family:'Helvetica Neue',Arial,sans-serif;font-size:16px;line-height:22px;text-align:center;color:#555;">
+                                        You have been invited to join <strong>{{ product }}</strong>.<br/><br/>
+                                        Click the invite link below to join.
+                                    </div>
+                                </td>
+                            </tr>
+
+                            <!-- Button -->
+                            <tr>
+                                <td align="center" style="padding:30px 25px 50px;">
+                                    <table align="center" border="0" cellpadding="0" cellspacing="0" role="presentation" style="border-collapse:separate;">
+                                        <tr>
+                                            <td align="center" bgcolor="#811429" role="presentation" style="border:none;border-radius:3px;padding:15px 25px;" valign="middle">
+                                                <a href="{{ inviteLink }}" style="font-family:'Helvetica Neue',Arial,sans-serif;font-size:15px;color:#ffffff;text-decoration:none;display:block;">
+                                                    Accept Invite
+                                                </a>
+                                            </td>
+                                        </tr>
+                                    </table>
+                                </td>
+                            </tr>
+
+                        </table>
+                    </td>
+                </tr></tbody>
+            </table>
+        </div>
+
+    </div>
+</body>
+</html>
